#26: mocked tests refactored
